Create new helper functions directory Move lottie and default step helpers into directory and export Rework lottie function so they run reliably
