+++ src/helperFunctions/runLottieAnimation.js
@@ -1,9 +1,10 @@
 import lottie from 'lottie-web';
 
-export default function runLottieAnimation(animationName, container) {
-    import(`./lotties/${animationName}.json`).then((animationData) => {
+export default function runLottieAnimation(animationName, containerId) {
+    console.log(document.getElementById(containerId))
+    import(`../lotties/${animationName}.json`).then((animationData) => {
         const animation = lottie.loadAnimation({
-            container: container,
+            container: document.getElementById(containerId),
             loop: false,
             autoplay: true,
             animationData: animationData,
